refactor: unify SQLAlchemy base
