+++ app/core/database.py
@@ -1,7 +1,6 @@
 # app/core/database.py
 
 from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
 from app.core.config import settings
@@ -9,6 +8,6 @@ from app.core.config import settings
 engine = create_engine(settings.database_url)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-Base = declarative_base()
+from app.db.base import Base
