+++ app/models/action.py
@@ -1,7 +1,6 @@
 from sqlalchemy import Column, Integer, Text, DateTime, ForeignKey
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
-from .base import Base
+from app.db.base import Base
 
 class Action(Base):
     __tablename__ = "actions"
