+++ app/models/bot.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
 from sqlalchemy.sql import func
-from .base import Base
+from app.db.base import Base
 
 class Bot(Base):
     __tablename__ = "bots"
