+++ app/models/channel.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, DateTime
 from sqlalchemy.sql import func
-from .base import Base
+from app.db.base import Base
 
 class Channel(Base):
     __tablename__ = "channels"
