+++ app/models/user.py
@@ -1,7 +1,6 @@
 from sqlalchemy import Column, Integer, String, DateTime, Boolean
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
-from .base import Base
+from app.db.base import Base
 
 class User(Base):
     __tablename__ = "users"
